requirements.txt
flask
flask_cors
groq
gunicorn
chromadb
python-dotenv
beautifulsoup4
requests
markdownify
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
open-clip-torch
